refactor: codeflare sdk unit tests
sort unit tests into individual files and made minor enhancements
