+++ src/codeflare_sdk/ray/cluster/test_generate_yaml.py
@@ -1,0 +1,34 @@
+# Copyright 2024 IBM, Red Hat
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from codeflare_sdk.ray.cluster.generate_yaml import gen_names
+import uuid
+
+
+def test_gen_names_with_name(mocker):
+    mocker.patch.object(
+        uuid, "uuid4", return_value=uuid.UUID("00000000-0000-0000-0000-000000000001")
+    )
+    name = "myname"
+    appwrapper_name, cluster_name = gen_names(name)
+    assert appwrapper_name == name
+    assert cluster_name == name
+
+
+def test_gen_names_without_name(mocker):
+    mocker.patch.object(
+        uuid, "uuid4", return_value=uuid.UUID("00000000-0000-0000-0000-000000000001")
+    )
+    appwrapper_name, cluster_name = gen_names(None)
+    assert appwrapper_name.startswith("appwrapper-")
+    assert cluster_name.startswith("cluster-")
