+++ src/codeflare_sdk/ray/cluster/test_pretty_print.py
@@ -1,0 +1,208 @@
+# Copyright 2024 IBM, Red Hat
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from codeflare_sdk.ray.cluster.pretty_print import (
+    print_app_wrappers_status,
+    print_cluster_status,
+    print_clusters,
+    print_no_resources_found,
+)
+from codeflare_sdk.ray.appwrapper.status import AppWrapperStatus, AppWrapper
+from codeflare_sdk.ray.cluster.status import (
+    RayCluster,
+    RayClusterStatus,
+    CodeFlareClusterStatus,
+)
+from codeflare_sdk.ray.cluster.cluster import (
+    Cluster,
+    ClusterConfiguration,
+    _copy_to_ray,
+)
+
+
+def test_print_no_resources(capsys):
+    try:
+        print_no_resources_found()
+    except Exception:
+        assert 1 == 0
+    captured = capsys.readouterr()
+    assert captured.out == (
+        "╭──────────────────────────────────────────────────────────────────────────────╮\n"
+        "│ No resources found, have you run cluster.up() yet?                           │\n"
+        "╰──────────────────────────────────────────────────────────────────────────────╯\n"
+    )
+
+
+def test_print_appwrappers(capsys):
+    aw1 = AppWrapper(
+        name="awtest1",
+        status=AppWrapperStatus.SUSPENDED,
+    )
+    aw2 = AppWrapper(
+        name="awtest2",
+        status=AppWrapperStatus.RUNNING,
+    )
+    try:
+        print_app_wrappers_status([aw1, aw2])
+    except Exception:
+        assert 1 == 0
+    captured = capsys.readouterr()
+    assert captured.out == (
+        "╭─────────────────────────╮\n"
+        "│     🚀 Cluster Queue    │\n"
+        "│        Status 🚀        │\n"
+        "│ +---------+-----------+ │\n"
+        "│ | Name    | Status    | │\n"
+        "│ +=========+===========+ │\n"
+        "│ | awtest1 | suspended | │\n"
+        "│ |         |           | │\n"
+        "│ | awtest2 | running   | │\n"
+        "│ |         |           | │\n"
+        "│ +---------+-----------+ │\n"
+        "╰─────────────────────────╯\n"
+    )
+
+
+def test_ray_details(mocker, capsys):
+    mocker.patch("kubernetes.client.ApisApi.get_api_versions")
+    ray1 = RayCluster(
+        name="raytest1",
+        status=RayClusterStatus.READY,
+        num_workers=1,
+        worker_mem_requests="2G",
+        worker_mem_limits="2G",
+        worker_cpu_requests=1,
+        worker_cpu_limits=1,
+        namespace="ns",
+        dashboard="fake-uri",
+        head_cpu_requests=2,
+        head_cpu_limits=2,
+        head_mem_requests=8,
+        head_mem_limits=8,
+    )
+    mocker.patch(
+        "codeflare_sdk.ray.cluster.cluster.Cluster.status",
+        return_value=(False, CodeFlareClusterStatus.UNKNOWN),
+    )
+    mocker.patch(
+        "codeflare_sdk.ray.cluster.cluster.Cluster.cluster_dashboard_uri",
+        return_value="",
+    )
+    mocker.patch(
+        "codeflare_sdk.common.kueue.kueue.local_queue_exists",
+        return_value="true",
+    )
+    cf = Cluster(
+        ClusterConfiguration(
+            name="raytest2",
+            namespace="ns",
+            appwrapper=True,
+            local_queue="local_default_queue",
+        )
+    )
+    captured = capsys.readouterr()
+    ray2 = _copy_to_ray(cf)
+    details = cf.details()
+    assert details == ray2
+    assert ray2.name == "raytest2"
+    assert ray1.namespace == ray2.namespace
+    assert ray1.num_workers == ray2.num_workers
+    assert ray1.worker_mem_requests == ray2.worker_mem_requests
+    assert ray1.worker_mem_limits == ray2.worker_mem_limits
+    assert ray1.worker_cpu_requests == ray2.worker_cpu_requests
+    assert ray1.worker_cpu_limits == ray2.worker_cpu_limits
+    assert ray1.worker_extended_resources == ray2.worker_extended_resources
+    try:
+        print_clusters([ray1, ray2])
+        print_cluster_status(ray1)
+        print_cluster_status(ray2)
+    except Exception:
+        assert 0 == 1
+    captured = capsys.readouterr()
+    assert captured.out == (
+        "                  🚀 CodeFlare Cluster Details 🚀                  \n"
+        "                                                                   \n"
+        " ╭───────────────────────────────────────────────────────────────╮ \n"
+        " │   Name                                                        │ \n"
+        " │   raytest2                                   Inactive ❌      │ \n"
+        " │                                                               │ \n"
+        " │   URI: ray://raytest2-head-svc.ns.svc:10001                   │ \n"
+        " │                                                               │ \n"
+        " │   Dashboard🔗                                                 │ \n"
+        " │                                                               │ \n"
+        " │                       Cluster Resources                       │ \n"
+        " │   ╭── Workers ──╮  ╭───────── Worker specs(each) ─────────╮   │ \n"
+        " │   │  # Workers  │  │  Memory      CPU         GPU         │   │ \n"
+        " │   │             │  │                                      │   │ \n"
+        " │   │  1          │  │  2G~2G       1~1         0           │   │ \n"
+        " │   │             │  │                                      │   │ \n"
+        " │   ╰─────────────╯  ╰──────────────────────────────────────╯   │ \n"
+        " ╰───────────────────────────────────────────────────────────────╯ \n"
+        "                  🚀 CodeFlare Cluster Details 🚀                  \n"
+        "                                                                   \n"
+        " ╭───────────────────────────────────────────────────────────────╮ \n"
+        " │   Name                                                        │ \n"
+        " │   raytest1                                   Active ✅        │ \n"
+        " │                                                               │ \n"
+        " │   URI: ray://raytest1-head-svc.ns.svc:10001                   │ \n"
+        " │                                                               │ \n"
+        " │   Dashboard🔗                                                 │ \n"
+        " │                                                               │ \n"
+        " │                       Cluster Resources                       │ \n"
+        " │   ╭── Workers ──╮  ╭───────── Worker specs(each) ─────────╮   │ \n"
+        " │   │  # Workers  │  │  Memory      CPU         GPU         │   │ \n"
+        " │   │             │  │                                      │   │ \n"
+        " │   │  1          │  │  2G~2G       1~1         0           │   │ \n"
+        " │   │             │  │                                      │   │ \n"
+        " │   ╰─────────────╯  ╰──────────────────────────────────────╯   │ \n"
+        " ╰───────────────────────────────────────────────────────────────╯ \n"
+        "╭───────────────────────────────────────────────────────────────╮\n"
+        "│   Name                                                        │\n"
+        "│   raytest2                                   Inactive ❌      │\n"
+        "│                                                               │\n"
+        "│   URI: ray://raytest2-head-svc.ns.svc:10001                   │\n"
+        "│                                                               │\n"
+        "│   Dashboard🔗                                                 │\n"
+        "│                                                               │\n"
+        "│                       Cluster Resources                       │\n"
+        "│   ╭── Workers ──╮  ╭───────── Worker specs(each) ─────────╮   │\n"
+        "│   │  # Workers  │  │  Memory      CPU         GPU         │   │\n"
+        "│   │             │  │                                      │   │\n"
+        "│   │  1          │  │  2G~2G       1~1         0           │   │\n"
+        "│   │             │  │                                      │   │\n"
+        "│   ╰─────────────╯  ╰──────────────────────────────────────╯   │\n"
+        "╰───────────────────────────────────────────────────────────────╯\n"
+        "                🚀 CodeFlare Cluster Status 🚀                \n"
+        "                                                              \n"
+        " ╭──────────────────────────────────────────────────────────╮ \n"
+        " │   Name                                                   │ \n"
+        " │   raytest1                                   Active ✅   │ \n"
+        " │                                                          │ \n"
+        " │   URI: ray://raytest1-head-svc.ns.svc:10001              │ \n"
+        " │                                                          │ \n"
+        " │   Dashboard🔗                                            │ \n"
+        " │                                                          │ \n"
+        " ╰──────────────────────────────────────────────────────────╯ \n"
+        "                 🚀 CodeFlare Cluster Status 🚀                 \n"
+        "                                                                \n"
+        " ╭────────────────────────────────────────────────────────────╮ \n"
+        " │   Name                                                     │ \n"
+        " │   raytest2                                   Inactive ❌   │ \n"
+        " │                                                            │ \n"
+        " │   URI: ray://raytest2-head-svc.ns.svc:10001                │ \n"
+        " │                                                            │ \n"
+        " │   Dashboard🔗                                              │ \n"
+        " │                                                            │ \n"
+        " ╰────────────────────────────────────────────────────────────╯ \n"
+    )
