+++ tests/test_cluster_yamls/ray/default-ray-cluster.yaml
@@ -1,14 +1,9 @@
 apiVersion: ray.io/v1
 kind: RayCluster
 metadata:
-  annotations:
-    app.kubernetes.io/managed-by: test-prefix
   labels:
     controller-tools.k8s.io: '1.0'
-    kueue.x-k8s.io/queue-name: local-queue-default
-    testlabel: test
-    testlabel2: test
-  name: unit-test-cluster-custom-image
+  name: default-cluster
   namespace: ns
 spec:
   autoscalerOptions:
@@ -34,7 +29,7 @@ spec:
     template:
       spec:
         containers:
-        - image: quay.io/project-codeflare/ray:2.20.0-py39-cu118
+        - image: quay.io/modh/ray@sha256:0d715f92570a2997381b7cafc0e224cfa25323f18b9545acfd23bc2b71576d06
           imagePullPolicy: Always
           lifecycle:
             preStop:
@@ -71,8 +66,7 @@ spec:
           - mountPath: /etc/ssl/certs/odh-ca-bundle.crt
             name: odh-ca-cert
             subPath: odh-ca-bundle.crt
-        imagePullSecrets:
-        - name: unit-test-pull-secret
+        imagePullSecrets: []
         volumes:
         - configMap:
             items:
@@ -90,14 +84,14 @@ spec:
           name: odh-ca-cert
   rayVersion: 2.35.0
   workerGroupSpecs:
-  - groupName: small-group-unit-test-cluster-custom-image
-    maxReplicas: 2
-    minReplicas: 2
+  - groupName: small-group-default-cluster
+    maxReplicas: 1
+    minReplicas: 1
     rayStartParams:
       block: 'true'
-      num-gpus: '7'
+      num-gpus: '0'
       resources: '"{}"'
-    replicas: 2
+    replicas: 1
     template:
       metadata:
         annotations:
@@ -106,7 +100,7 @@ spec:
           key: value
       spec:
         containers:
-        - image: quay.io/project-codeflare/ray:2.20.0-py39-cu118
+        - image: quay.io/modh/ray@sha256:0d715f92570a2997381b7cafc0e224cfa25323f18b9545acfd23bc2b71576d06
           lifecycle:
             preStop:
               exec:
@@ -117,13 +111,11 @@ spec:
           name: machine-learning
           resources:
             limits:
-              cpu: 4
-              memory: 6G
-              nvidia.com/gpu: 7
+              cpu: 1
+              memory: 2G
             requests:
-              cpu: 3
-              memory: 5G
-              nvidia.com/gpu: 7
+              cpu: 1
+              memory: 2G
           volumeMounts:
           - mountPath: /etc/pki/tls/certs/odh-trusted-ca-bundle.crt
             name: odh-trusted-ca-cert
@@ -137,8 +129,7 @@ spec:
           - mountPath: /etc/ssl/certs/odh-ca-bundle.crt
             name: odh-ca-cert
             subPath: odh-ca-bundle.crt
-        imagePullSecrets:
-        - name: unit-test-pull-secret
+        imagePullSecrets: []
         volumes:
         - configMap:
             items:
